Speed up rendering whole life chart
Profiling showed that we were repeatedly calculating maxX and minX, even though they don't change for a given render.

Moved them to be calculated once up front and then passed down to the other calculations.

I also changed them from a `get` to a method to make it clearer that they do "heavy" calculation (they're not that heavy but if you do a whole bunch of them in a row it gets slow!)